api service order complete finishhh

+++ lib/dashboard/model/dashboard_service.dart
@@ -121,7 +121,7 @@ class OrderType {
 }
 
 class NewOrder {
-  int? values;
+  double? values;
   int? items;
 
   NewOrder({this.values, this.items});
